Fix bug in diagonal_move()
In the 2-knot rope, the head always moves in a cardinal direction (U/D/L/R), and this method computes the tail movement correctly in that case.

However, in the 10-knot rope, we use this same method to calculate movement for all 9 "tails", and some of the preceding "heads" might have already moved diagonally. In the case where `head` has already moved diagonally away from `tail` (e.g. `head` moved `(1, 1) -> (2, 2)`, and we're now moving `tail` from `(0, 0)`) the old code would make move `tail` skip to `(1, 2)` instead of following diagonally to `(1, 1)`.